商品削除機能の修正

+++ app/controllers/items_controller.rb
@@ -24,8 +24,12 @@ class ItemsController < ApplicationController
 
   def destroy
     item = Item.find(params[:id])
-    item.destroy
-    redirect_to root_path
+    if item.valid?
+      item.destroy
+      redirect_to root_path
+    else
+      render action: :show
+    end
   end
 
   private
